mica/requirements.txt
ultralytics
opencv-python
numpy
pyyaml
requests
torch
torchvision
scikit-learn
# Optional for gallery embed=clip:
# open-clip-torch
# Pillow
